Add test for Filter Component
